Getting rid of math.h
